Change FPDFFont_GetFamilyName() to take/return a size_t
Many other APIs in the same header are using size_t already, so switch
to the type that better represents buffer size. Then the implementation
can do 1 fewer cast.

Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/121894
Commit-Queue: Lei Zhang <thestig@chromium.org>

+++ fpdfsdk/fpdf_edittext.cpp
@@ -849,8 +849,9 @@ FPDF_EXPORT FPDF_FONT FPDF_CALLCONV FPDFTextObj_GetFont(FPDF_PAGEOBJECT text) {
   return FPDFFontFromCPDFFont(pTextObj->GetFont());
 }
 
-FPDF_EXPORT unsigned long FPDF_CALLCONV
-FPDFFont_GetFamilyName(FPDF_FONT font, char* buffer, unsigned long length) {
+FPDF_EXPORT size_t FPDF_CALLCONV FPDFFont_GetFamilyName(FPDF_FONT font,
+                                                        char* buffer,
+                                                        size_t length) {
   auto* cfont = CPDFFontFromFPDFFont(font);
   if (!cfont) {
     return 0;
@@ -861,7 +862,7 @@ FPDFFont_GetFamilyName(FPDF_FONT font, char* buffer, unsigned long length) {
   ByteString name = cfont->GetFont()->GetFamilyName();
   pdfium::span<const char> name_span = name.span_with_terminator();
   fxcrt::try_spancpy(result_span, name_span);
-  return static_cast<unsigned long>(name_span.size());
+  return name_span.size();
 }
 
 FPDF_EXPORT FPDF_BOOL FPDF_CALLCONV FPDFFont_GetFontData(FPDF_FONT font,
